ENH: move _partition_estimators to base.py
