+++ sklearn/ensemble/forest.py
@@ -44,7 +44,7 @@ from warnings import warn
 from abc import ABCMeta, abstractmethod
 
 from ..base import ClassifierMixin, RegressorMixin
-from ..externals.joblib import Parallel, delayed, cpu_count
+from ..externals.joblib import Parallel, delayed
 from ..externals import six
 from ..externals.six.moves import xrange
 from ..feature_selection.from_model import _LearntSelectorMixin
@@ -57,8 +57,7 @@ from ..utils import array2d, check_random_state, check_arrays, safe_asarray
 from ..utils.validation import DataConversionWarning
 from ..utils.fixes import bincount, unique
 
-
-from .base import BaseEnsemble
+from .base import BaseEnsemble, _partition_estimators
 
 __all__ = ["RandomForestClassifier",
            "RandomForestRegressor",
@@ -149,23 +148,6 @@ def _parallel_predict_proba(trees, X, n_classes, n_outputs):
 def _parallel_predict_regression(trees, X):
     """Private function used to compute a batch of predictions within a job."""
     return sum(tree.predict(X) for tree in trees)
-
-
-def _partition_trees(forest):
-    """Private function used to partition trees between jobs."""
-    # Compute the number of jobs
-    if forest.n_jobs == -1:
-        n_jobs = min(cpu_count(), forest.n_estimators)
-
-    else:
-        n_jobs = min(forest.n_jobs, forest.n_estimators)
-
-    # Partition trees between jobs
-    n_trees = (forest.n_estimators // n_jobs) * np.ones(n_jobs, dtype=np.int)
-    n_trees[:forest.n_estimators % n_jobs] += 1
-    starts = np.cumsum(n_trees)
-
-    return n_jobs, n_trees.tolist(), [0] + starts.tolist()
 
 
 class BaseForest(six.with_metaclass(ABCMeta, BaseEnsemble,
@@ -274,7 +256,7 @@ class BaseForest(six.with_metaclass(ABCMeta, BaseEnsemble,
                              " if bootstrap=True")
 
         # Assign chunk of trees to jobs
-        n_jobs, n_trees, _ = _partition_trees(self)
+        n_jobs, n_trees, _ = _partition_estimators(self)
 
         # Precalculate the random states
         seeds = [random_state.randint(MAX_INT, size=i) for i in n_trees]
@@ -469,7 +451,7 @@ class ForestClassifier(six.with_metaclass(ABCMeta, BaseForest,
             X = array2d(X, dtype=DTYPE)
 
         # Assign chunk of trees to jobs
-        n_jobs, n_trees, starts = _partition_trees(self)
+        n_jobs, n_trees, starts = _partition_estimators(self)
 
         # Parallel loop
         all_proba = Parallel(n_jobs=n_jobs, verbose=self.verbose)(
@@ -578,7 +560,7 @@ class ForestRegressor(six.with_metaclass(ABCMeta, BaseForest, RegressorMixin)):
             X = array2d(X, dtype=DTYPE)
 
         # Assign chunk of trees to jobs
-        n_jobs, n_trees, starts = _partition_trees(self)
+        n_jobs, n_trees, starts = _partition_estimators(self)
 
         # Parallel loop
         all_y_hat = Parallel(n_jobs=n_jobs, verbose=self.verbose)(
